# server/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "### Fast weblog built in golang and top of echo. supports React SSR and hydrate.\nlabstack/echoを使った高速なブログエンジン。\nReactのサーバーサイドレンダリングとクライアントサイドの引き継ぎを実装。\n（ぶっちゃけ速いかと言われるとSSR有効時はそうでもない）",
        "title": "parakeet REST API",
        "contact": {},
        "license": {
            "name": "WTFPL",
            "url": "https://github.com/mohemohe/parakeet/blob/master/LICENSE"
        },
        "version": "1.0 Draft"
    },
    "basePath": "/api",
    "paths": {
...
        "/v1/drive/{path}": {
            "get": {
                "description": "ファイル一覧またはファイルを取得します",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "drive"
                ],
                "summary": "list files",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ドライブのパス",
                        "name": "path",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/s3fs.FileInfo"
                            }
                        }
                    }
                }
            },
            "put": {
                "description": "ファイルを移動またはコピーします",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "drive"
                ],
                "summary": "list files",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ドライブのコピー先パス",
                        "name": "path",
                        "in": "query",
                        "required": true
                    },
                    {
                        "description": "Body",
                        "name": "Body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controllers.DriveFileCopyRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                }
            },
            "post": {
                "description": "ファイルを追加またはフォルダーを作成します",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "drive"
                ],
                "summary": "list files",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ドライブの作成先パス",
                        "name": "path",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                }
            },
            "delete": {
                "description": "ファイルを削除します",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "drive"
                ],
                "summary": "delete file(s)",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ファイル パス",
                        "name": "path",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "controllers.DriveFileCopyRequest": {
            "type": "object",
            "properties": {
                "operation": {
                    "type": "string"
                },
                "src": {
                    "type": "string"
                }
            }
        },
...
        "s3fs.FileInfo": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                },
                "path": {
                    "type": "string"
                },
                "size": {
                    "type": "integer"
                },
                "type": {
                    "type": "integer"
                }
            }
        }
    },
    "securityDefinitions": {
        "AccessToken": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}